consts to usize for platform dependent sizing

+++ src/platform/i2c.rs
@@ -4,15 +4,15 @@ use ioctl_consts::*;
 
 #[allow(dead_code)]
 mod ioctl_consts {
-    pub const I2C_RETRIES: u64 = 0x0701;
-    pub const I2C_TIMEOUT: u64 = 0x0702;
-    pub const I2C_SLAVE: u64 = 0x0703;
-    pub const I2C_SLAVE_FORCE: u64 = 0x0706;
-    pub const I2C_TENBIT: u64 = 0x0704;
-    pub const I2C_FUNCS: u64 = 0x0705;
-    pub const I2C_RDWR: u64 = 0x0707;
-    pub const I2C_PEC: u64 = 0x0708;
-    pub const I2C_SMBUS: u64 = 0x0720;
+    pub const I2C_RETRIES: usize = 0x0701;
+    pub const I2C_TIMEOUT: usize = 0x0702;
+    pub const I2C_SLAVE: usize = 0x0703;
+    pub const I2C_SLAVE_FORCE: usize = 0x0706;
+    pub const I2C_TENBIT: usize = 0x0704;
+    pub const I2C_FUNCS: usize = 0x0705;
+    pub const I2C_RDWR: usize = 0x0707;
+    pub const I2C_PEC: usize = 0x0708;
+    pub const I2C_SMBUS: usize = 0x0720;
 }
 
 /*
